requirements/development.txt
# Import common.txt packages
-r common.txt

# Data loading and processing
pandas==0.22.0
xlrd==1.1.0
django-autoslug-iplweb==1.9.4.dev0
